skipping if we already downloaded the file

+++ downloader.rb
@@ -15,18 +15,20 @@ class Downloader
     pages.each do |page|
       browser.visit("#{BASE_URL}?page=#{page}")
       browser.find_all(".pretty_button").each.with_index do |a_tag, i|
-        video_links << a_tag[:href] if i.odd?
+        video_links << a_tag[:href]
       end
     end
 
     video_links.each do |video_link|
-      file_name = browser.title.gsub(" ", "_")
+      file_name = browser.title.gsub(" ", "_").gsub(/[^\w]/, "")
       browser.visit(video_link)
       browser.click_link("mp4")
       url = browser.current_url
       browser.go_back
 
       file_path = "#{DOWNLOAD_FOLDER}#{file_name}.mp4"
+      next if File.exist?(file_path)
+
       open(file_path, "wb") do |file|
         file.print open(url).read
       end
